requirements.txt
dashscope
pydantic
python-dotenv
pdfplumber
pytesseract
tenacity
Pillow
